Code review fixes

+++ app/src/main/java/com/artemissoftware/tasky/agenda/domain/models/Notification.kt
@@ -1,6 +1,6 @@
 package com.artemissoftware.tasky.agenda.domain.models
 
 data class Notification(
-    val minutesBefore: Long,
+    val minutesBefore: Int,
     val description: String,
 )
